// src/app/models/product.ts
export interface Product {
  // id: any;
  // name: string;
  // price: number;
  // imageUrl: string;
  // description: string;
  // isActive: boolean;
  // categoryId: number;
  id: any;
  name: string;
  userId: string;
  adImages: string;
  adNo: string;
  adDate: string;
  categoryId: number;
  phoneNo: string;
  priceFree: string;
  adDescription: string;
  city: string;
  district: string;
  quarter: string;
  fullAddress: string;
  latitude: string;
  longitude: string;
  isActive: boolean;
}